up to now(2018.02.24)

+++ scripts/install_prerequisites.sh
@@ -25,12 +25,18 @@
 #
 # Note that the root folder MUST be named 'oppia'.
 
+if [ -e "/etc/is_vagrant_vm" ]
+then
+  source $(dirname $0)/vagrant_lock.sh || exit 1
+fi
+
 sudo apt-get update
 sudo apt-get install curl
 sudo apt-get install git
-sudo apt-get install openjdk-7-jre
+sudo apt-get install openjdk-8-jre
 sudo apt-get install python-setuptools
 sudo apt-get install python-dev
 sudo apt-get install python-pip
 sudo apt-get install unzip
+sudo apt-get install python-yaml
 sudo pip install --upgrade pip
